fix: Browser page titles
