+++ content/en/blog.md
@@ -1,3 +1,8 @@
+---
+title: Blog
+show_title: false
+---
+
 Find out more about what the Preact team is working on.
 
 <div><blog-overview></blog-overview></div>
