+++ content/ru/blog.md
@@ -1,3 +1,8 @@
+---
+title: Blog
+show_title: false
+---
+
 Узнайте больше о том, над чем работает команда Preact.
 
 <div><blog-overview></blog-overview></div>
